191112| TIL

+++ 05_Django/05_django_form/articles/views.py
@@ -27,7 +27,9 @@ def create(request):
             # title = form.cleaned_data.get('title')
             # content = form.cleaned_data.get('content')
             # article = Article.objects.create(title=title, content=content)
-            article = form.save()
+            article = form.save(commit=False)
+            article.user = request.user
+            article.save()
         return redirect('articles:detail', article.pk)
     else:
         form = ArticleForm()
@@ -51,32 +53,42 @@ def detail(request, article_pk):
 
 @require_POST
 def delete(request, article_pk):
+    # 지금 사용자가 로그인 되어있는지?
     if request.user.is_authenticated:
+        # 삭제할 게시글 가져오기
         article = get_object_or_404(Article, pk=article_pk)
-        article.delete()
+        # 로그인한 사용자와 게시글 작성자 비교
+        if request.user == article.user:
+            article.delete()
+        else:
+            return redirect('articles:detail', article.pk)
     return redirect('articles:index')
 
 @login_required
 def update(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
-    if request.method == 'POST':
-        # form = ArticleForm(request.POST)
-        form = ArticleForm(request.POST, instance=article)
-        if form.is_valid():
-            # article.title = form.cleaned_data.get('title')
-            # article.content = form.cleaned_data.get('content')
-            # article.save()
-            article=form.save()
-            return redirect('articles:detail', article.pk)
+    if request.user == article.user:
+        if request.method == 'POST':
+            # form = ArticleForm(request.POST)
+            form = ArticleForm(request.POST, instance=article)
+            if form.is_valid():
+                # article.title = form.cleaned_data.get('title')
+                # article.content = form.cleaned_data.get('content')
+                # article.save()
+                article=form.save()
+                return redirect('articles:detail', article.pk)
+        else:
+            # form = ArticleForm(initial={
+            #     'title' : article.title,
+            #     'content' : article.content
+            # })
+            form = ArticleForm(instance=article)
     else:
-        # form = ArticleForm(initial={
-        #     'title' : article.title,
-        #     'content' : article.content
-        # })
-        form = ArticleForm(instance=article)
-        # context로 전달되는 2가지 form 형식
-        # 1. GET -> 초기값을 펌에 넣어서 사용자에게 던져줌
-        # 2. POST -> is_valid가 False가 리턴되었을 때, 오류메시지 포함해서 사용자에게 던져줌
+        return redirect('articles:index')
+    
+    # context로 전달되는 2가지 form 형식
+    # 1. GET -> 초기값을 펌에 넣어서 사용자에게 던져줌
+    # 2. POST -> is_valid가 False가 리턴되었을 때, 오류메시지 포함해서 사용자에게 던져줌
     context = {
         'form' : form,
         'article' : article,    
@@ -94,13 +106,17 @@ def comments_create(request, article_pk):
             # DB에 바로 저장되는 것을 막아준다.
             comment = comment_form.save(commit=False)
             comment.article = article
+            comment.user = article.user
             comment.save()
     return redirect('articles:detail', article.pk)
     
 
 @require_POST
 def comments_delete(request, article_pk, comment_pk):
+    # 1. 로그인 여부 확인
     if request.user.is_authenticated:
         comment = get_object_or_404(Comment, pk=comment_pk)
-        comment.delete()
+        # 2. 로그인한 사용자와 댓글 작성자가 같을 경우
+        if request.user == comment.user:
+            comment.delete()
     return redirect('articles:detail', article_pk)
